+++ makeflow/src/dag_visitors.c
@@ -1184,8 +1184,8 @@ struct jx *category_allocation_to_json(category_allocation_t c) {
 		case CATEGORY_ALLOCATION_ERROR:
 			return jx_string("error");
 			break;
-		case CATEGORY_ALLOCATION_FAST_BUCKETING:
-			return jx_string("fast bucketing");
+		case CATEGORY_ALLOCATION_GREEDY_BUCKETING:
+			return jx_string("greedy bucketing");
 			break;
 		case CATEGORY_ALLOCATION_EXHAUSTIVE_BUCKETING:
 			return jx_string("exhaustive bucketing");
