+++ parrot/src/.gitignore
@@ -15,6 +15,8 @@ parrot_search
 parrot_setacl
 parrot_timeout
 parrot_whoami
+parrot_test_dir
+parrot_test_execve
 tracer.table.c
 tracer.table.h
 tracer.table64.c
