+++ work_queue/src/work_queue.c
@@ -7833,8 +7833,8 @@ static void write_transaction_category(struct work_queue *q, struct category *c)
 		default:
 			mode = "FIXED";
 			break;
-		case CATEGORY_ALLOCATION_MODE_FAST_BUCKETING:
-			mode = "FAST_BUCKETING";
+		case CATEGORY_ALLOCATION_MODE_GREEDY_BUCKETING:
+			mode = "GREEDY_BUCKETING";
 			break;
 		case CATEGORY_ALLOCATION_MODE_EXHAUSTIVE_BUCKETING:
 			mode = "EXHAUSTIVE_BUCKETING";
@@ -8056,7 +8056,7 @@ int work_queue_specify_category_mode(struct work_queue *q, const char *category,
 		case WORK_QUEUE_ALLOCATION_MODE_MAX:
 		case WORK_QUEUE_ALLOCATION_MODE_MIN_WASTE:
 		case WORK_QUEUE_ALLOCATION_MODE_MAX_THROUGHPUT:
-		case WORK_QUEUE_ALLOCATION_MODE_FAST_BUCKETING:
+		case WORK_QUEUE_ALLOCATION_MODE_GREEDY_BUCKETING:
 		case WORK_QUEUE_ALLOCATION_MODE_EXHAUSTIVE_BUCKETING:
 			break;
 		default:
